requirements.txt
google-cloud-storage
langchain_core
langchain_experimental
langchain
docarray==0.39.1
pydantic==1.10.9
pyhcl
google-generativeai==0.3.1
google-ai-generativelanguage==0.4.0
google-cloud-aiplatform
python-dotenv
google-search-results
google-auth
pypdf
unstructured
tiktoken
pinecone-client
cohere
jq
openai
langchain-google-genai
slack-sdk
slack-bolt